+++ cmd/pipeline/migrate.go
@@ -22,7 +22,7 @@ import (
 	"github.com/banzaicloud/pipeline/internal/cluster/clusteradapter"
 	"github.com/banzaicloud/pipeline/internal/clustergroup/deployment"
 	"github.com/banzaicloud/pipeline/internal/common"
-	"github.com/banzaicloud/pipeline/internal/helm3"
+	"github.com/banzaicloud/pipeline/internal/helm3/helmrepoadapter"
 	"github.com/banzaicloud/pipeline/internal/integratedservices/integratedserviceadapter"
 
 	"github.com/banzaicloud/pipeline/internal/app/pipeline/api/middleware/audit"
@@ -85,7 +85,7 @@ func Migrate(db *gorm.DB, logger logrus.FieldLogger, commonLogger common.Logger)
 		return err
 	}
 
-	if err := helm3.Migrate(db, logger); err != nil {
+	if err := helmrepoadapter.Migrate(db, logger); err != nil {
 		return err
 	}
