+++ internal/helm3/helmrepoadapter/gorm_repo_store.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helm3
+package helmrepoadapter
 
 import (
 	"context"
@@ -21,6 +21,7 @@ import (
 	"github.com/jinzhu/gorm"
 
 	"github.com/banzaicloud/pipeline/internal/common"
+	"github.com/banzaicloud/pipeline/internal/helm3"
 )
 
 type helmRepoStore struct {
@@ -28,14 +29,14 @@ type helmRepoStore struct {
 	logger common.Logger
 }
 
-func NewHelmRepoStore(db *gorm.DB, logger common.Logger) Store {
+func NewHelmRepoStore(db *gorm.DB, logger common.Logger) helm3.Store {
 	return helmRepoStore{
 		db:     db,
 		logger: logger,
 	}
 }
 
-func (h helmRepoStore) DeleteRepository(_ context.Context, organizationID uint, repository Repository) error {
+func (h helmRepoStore) DeleteRepository(_ context.Context, organizationID uint, repository helm3.Repository) error {
 	model := ToModel(repository)
 	model.OrganizationID = organizationID
 
@@ -53,13 +54,13 @@ func (h helmRepoStore) DeleteRepository(_ context.Context, organizationID uint, 
 	return nil
 }
 
-func (h helmRepoStore) ListRepositories(_ context.Context, organizationID uint) ([]Repository, error) {
+func (h helmRepoStore) ListRepositories(_ context.Context, organizationID uint) ([]helm3.Repository, error) {
 	var repoModels []RepositoryModel
 	if err := h.db.Where("organization_id = ?", organizationID).Find(&repoModels).Error; err != nil {
 		return nil, errors.WrapIf(err, "failed to list helm repositories")
 	}
 
-	repos := make([]Repository, 0, len(repoModels))
+	repos := make([]helm3.Repository, 0, len(repoModels))
 	for _, model := range repoModels {
 		repos = append(repos, ToDomain(model))
 	}
@@ -70,7 +71,7 @@ func (h helmRepoStore) ListRepositories(_ context.Context, organizationID uint) 
 	return repos, nil
 }
 
-func (h helmRepoStore) AddRepository(_ context.Context, organizationID uint, repository Repository) error {
+func (h helmRepoStore) AddRepository(_ context.Context, organizationID uint, repository helm3.Repository) error {
 	repoModel := ToModel(repository)
 	repoModel.OrganizationID = organizationID
 
@@ -83,9 +84,22 @@ func (h helmRepoStore) AddRepository(_ context.Context, organizationID uint, rep
 	return nil
 }
 
+func (h helmRepoStore) GetRepository(_ context.Context, organizationID uint, repository helm3.Repository) (helm3.Repository, error) {
+	repoModel := ToModel(repository)
+	repoModel.OrganizationID = organizationID
+
+	if err := h.db.Where(&repoModel).First(&repoModel).Error; err != nil {
+		return helm3.Repository{}, errors.WrapIf(err, "failed to list helm repositories")
+	}
+	h.logger.Debug("retrieved helm repository record",
+		map[string]interface{}{"organisationID": organizationID, "repoName": repository.Name})
+
+	return ToDomain(repoModel), nil
+}
+
 // ToDomain transforms a gorm model to a domain struct
-func ToDomain(model RepositoryModel) Repository {
-	return Repository{
+func ToDomain(model RepositoryModel) helm3.Repository {
+	return helm3.Repository{
 		Name:             model.Name,
 		URL:              model.URL,
 		PasswordSecretID: model.PasswordSecretID,
@@ -94,7 +108,7 @@ func ToDomain(model RepositoryModel) Repository {
 }
 
 //ToModel transforms a domain struct to gorm model representation
-func ToModel(repository Repository) RepositoryModel {
+func ToModel(repository helm3.Repository) RepositoryModel {
 	return RepositoryModel{
 		Name:             repository.Name,
 		PasswordSecretID: repository.PasswordSecretID,
